More pr comments and test refactor
