+++ src/test/scala/io/iohk/ethereum/vm/OpCodeGasSpec.scala
@@ -1,16 +1,14 @@
 package io.iohk.ethereum.vm
 
-import org.scalatest.{FunSuite, Matchers}
-import org.scalatest.prop.PropertyChecks
-import Generators._
-import io.iohk.ethereum.domain.{Account, Address}
-import io.iohk.ethereum.domain.UInt256
 import io.iohk.ethereum.domain.UInt256._
-import io.iohk.ethereum.vm.MockWorldState.PS
+import io.iohk.ethereum.domain.{Account, Address, UInt256}
+import io.iohk.ethereum.vm.Generators._
+import org.scalatest.prop.PropertyChecks
+import org.scalatest.{FunSuite, Matchers}
 
 class OpCodeGasSpec extends FunSuite with OpCodeTesting with Matchers with PropertyChecks {
 
-  override val config = EvmConfig.PostEIP161ConfigBuilder(None)
+  override val config = EvmConfig.PostEIP160ConfigBuilder(None)
 
   import config.feeSchedule._
 
@@ -60,18 +58,6 @@ class OpCodeGasSpec extends FunSuite with OpCodeTesting with Matchers with Prope
     JUMPDEST -> G_jumpdest
   ) ++ stackOpsFees ++ constOpsFees
 
-  def verifyGas(expectedGas: BigInt, stateIn: PS, stateOut: PS, allowOOG: Boolean = true): Unit = {
-    if (stateOut.error.contains(OutOfGas) && allowOOG)
-      stateIn.gas should be < expectedGas
-    else if (stateOut.error.contains(OutOfGas) && !allowOOG)
-      fail(s"Unexpected $OutOfGas error")
-    else if (stateOut.error.isDefined && stateOut.error.collect{ case InvalidJump(dest) => dest }.isEmpty)
-      //Found error that is not an InvalidJump
-      fail(s"Unexpected ${stateOut.error.get} error")
-    else {
-      stateOut.gas shouldEqual (stateIn.gas - expectedGas)
-    }
-  }
 
   test("wordsForBytes helper") {
     val testData = Table(("bytes", "words"),
@@ -477,19 +463,19 @@ class OpCodeGasSpec extends FunSuite with OpCodeTesting with Matchers with Prope
     // Sending refund to a non-existent account
     forAll(stateGen) { stateIn =>
       val (refund, _) = stateIn.stack.pop
-      whenever(stateIn.world.getAccount(Address(refund)).isEmpty && stateIn.ownBalance > 0) {
+      whenever(stateIn.world.getAccount(Address(refund)).isEmpty) {
         val stateOut = op.execute(stateIn)
         stateOut.gasRefund shouldEqual R_selfdestruct
         verifyGas(G_selfdestruct + G_newaccount, stateIn, stateOut)
       }
     }
 
-    // Sending refund to an already existing account not dead account
+    // Sending refund to an already existing account
     forAll(stateGen) { (stateIn) =>
       val (refund, _) = stateIn.stack.pop
       val world = stateIn.world.saveAccount(
         Address(refund),
-        Account.empty().increaseNonce())
+        Account.empty())
       val updatedStateIn = stateIn.withWorld(world)
       val stateOut = op.execute(updatedStateIn)
       verifyGas(G_selfdestruct, updatedStateIn, stateOut)
@@ -499,13 +485,14 @@ class OpCodeGasSpec extends FunSuite with OpCodeTesting with Matchers with Prope
     // Owner account was already selfdestructed
     forAll(stateGen) { stateIn =>
       val (refund, _) = stateIn.stack.pop
-      whenever(stateIn.world.getAccount(Address(refund)).isEmpty && stateIn.ownBalance > 0) {
+      whenever(stateIn.world.getAccount(Address(refund)).isEmpty) {
         val updatedStateIn = stateIn.withAddressToDelete(stateIn.context.env.ownerAddr)
         val stateOut = op.execute(updatedStateIn)
         verifyGas(G_selfdestruct + G_newaccount, updatedStateIn, stateOut)
         stateOut.gasRefund shouldEqual 0
       }
     }
+
   }
 
   verifyAllOpCodesRegistered(except = CREATE, CALL, CALLCODE, DELEGATECALL, INVALID)
